Import Autorizacion Cobro

+++ app/src/main/java/union/union_vr1/AsyncTask/ExportMain.java
@@ -521,7 +521,7 @@ public class ExportMain extends AsyncTask<String, String, String> {
                     jsonObject = api.CreateInsertarCaja(
                             ((MyApplication)mainActivity.getApplication()).getIdLiquidacion(),
                             2,
-                            cursorInsertarCaja.getInt(cursorInsertarCaja.getColumnIndexOrThrow(dbAdapter_comprob_cobro.CC_monto_cobrado)),
+                            cursorInsertarCaja.getDouble(cursorInsertarCaja.getColumnIndexOrThrow(dbAdapter_comprob_cobro.CC_monto_cobrado)),
                             cursorInsertarCaja.getInt(cursorInsertarCaja.getColumnIndexOrThrow(dbAdapter_comprob_cobro.CC_estado_cobro)),
                             cursorInsertarCaja.getString(cursorInsertarCaja.getColumnIndexOrThrow(dbAdapter_comprob_cobro.CC_fecha_cobro)),
                             String.valueOf(cursorInsertarCaja.getInt(cursorInsertarCaja.getColumnIndexOrThrow(dbAdapter_comprob_cobro.CC_id_comprobante_cobro))),
